// src/viteplus-monorepo/apps/console/src/__generated/projects-api/valibot.gen.ts
export const vCreateProjectRequest = v.strictObject({
  $schema: v.optional(v.pipe(v.pipe(v.string(), v.url()), v.readonly())),
  description: v.optional(v.pipe(v.string(), v.maxLength(500))),
  display_name: v.pipe(v.string(), v.minLength(1), v.maxLength(120)),
  slug: v.optional(v.pipe(v.string(), v.minLength(1), v.maxLength(80))),
});
// ...
export const vProject = v.strictObject({
  $schema: v.optional(v.pipe(v.pipe(v.string(), v.url()), v.readonly())),
  archived_at: v.optional(v.pipe(v.string(), v.isoTimestamp())),
  created_at: v.pipe(v.string(), v.isoTimestamp()),
  created_by: v.string(),
  description: v.string(),
  display_name: v.string(),
  org_id: v.pipe(v.string(), v.regex(/^[0-9]+$/)),
  project_id: v.pipe(v.string(), v.uuid()),
  slug: v.string(),
  state: v.string(),
  updated_at: v.pipe(v.string(), v.isoTimestamp()),
  updated_by: v.string(),
  version: v.pipe(v.string(), v.regex(/^-?[0-9]+$/)),
});
// ...
export const vProjectList = v.strictObject({
  $schema: v.optional(v.pipe(v.pipe(v.string(), v.url()), v.readonly())),
  next_cursor: v.optional(v.string()),
  projects: v.nullable(v.array(vProject)),
});
// ...
export const vCreateProjectRequestWritable = v.strictObject({
  description: v.optional(v.pipe(v.string(), v.maxLength(500))),
  display_name: v.pipe(v.string(), v.minLength(1), v.maxLength(120)),
  slug: v.optional(v.pipe(v.string(), v.minLength(1), v.maxLength(80))),
});
// ...
export const vListProjectsQuery = v.object({
  state: v.optional(v.picklist(["active", "archived"])),
  limit: v.optional(
    v.pipe(
      v.union([v.number(), v.string(), v.bigint()]),
      v.transform((x) => BigInt(x)),
      v.minValue(BigInt(1)),
      v.maxValue(BigInt(100)),
    ),
  ),
  cursor: v.optional(v.pipe(v.string(), v.maxLength(256))),
});

/**
 * OK
 */
export const vListProjectsResponse = vProjectList;

export const vCreateProjectBody = vCreateProjectRequestWritable;

export const vCreateProjectHeaders = v.object({
  "Idempotency-Key": v.pipe(v.string(), v.minLength(1), v.maxLength(128)),
});

/**
 * Created
 */
export const vCreateProjectResponse = vProject;

export const vGetProjectPath = v.object({
  project_id: v.pipe(v.string(), v.uuid()),
});